refactor(consensus): Split consensus.py in multiple files
